+++ examples/advanced/kubb.config.ts
@@ -22,7 +22,7 @@ export default defineConfig(async () => {
     plugins: [
       ['@kubb/swagger', { output: false, validate: true }],
       ['@kubb/swagger-ts', { output: 'models/ts' }],
-      ['@kubb/swagger-react-query', { output: './reactQuery', groupBy: 'tag', client: './src/client.ts' }],
+      ['@kubb/swagger-tanstack-query', { output: './reactQuery', groupBy: 'tag', client: './src/client.ts' }],
       ['@kubb/swagger-client', { output: './clients', groupBy: 'tag', client: './src/client.ts' }],
       ['@kubb/swagger-zod', { output: './zod' }],
       // createSwagger({ output: false }),