+++ examples/advanced/src/client.ts
@@ -1,6 +1,6 @@
 import axios from 'axios'
 
-import type { RequestConfig } from '@kubb/swagger-react-query'
+import type { RequestConfig } from '@kubb/swagger-client'
 
 export const axiosClient = async <TData, TVariables = unknown>(config: RequestConfig<TVariables>) => {
   const promise = axios.request<TData>({ ...config }).then(({ data }) => data)