+++ examples/simple-single/kubb.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from '@kubb/core'
 import createSwagger from '@kubb/swagger'
 import createSwaggerTS from '@kubb/swagger-ts'
-import createSwaggerReactQuery from '@kubb/swagger-react-query'
+import createSwaggerReactQuery from '@kubb/swagger-tanstack-query'
 
 export default defineConfig({
   root: '.',